Dependabot skips Image Scans
As Dependabot triggered builds only have read credentials for the Docker
Hub registry they already don't push the built images.  However, this
means that when the scan image workflow is triggered it tries to pull a
non-existent image and fails.

Thus we now skip the image scan workflow for Dependabot builds.