Auto-create Maven Release PRs
Adds a new step to the Maven GitHub Release workflow so that on
successful release it will create a Release PR that completes the
release process.  This avoids the need for developers to manually
open said PRs and reduces the chance that these get missed.

Note no reviewers are assigned, rather we'll rely on default reviewers
which are populated from CODEOWNERS if present