Make opening Maven Release PR not break build
As seen in SCG build if the developer is proactive in creating the
Release PR, or it already exists for some other reason, then this breaks
the build and causes artefacts to not be correctly published.  Make the
Release PR step continue-on-error: true so that failures in that step
are ignored.