Telfe 211/merger (#61)
* Split reusable Maven workflow pieces into own workflows

* Add parallel Maven workflow

Adds a parallel variant of our Maven workflow

* Document new parallel Maven workflow

Plus some documentation clarifications and corrections

* Inherit secrets between Maven workflow components

Co-authored-by: Greg-teli100 <143502364+Greg-teli100@users.noreply.github.com>
Signed-off-by: Rob Vesse <rob.vesse@telicent.io>

* Pin called workflows to main

* Add support for Grype image scanning (CORE-440)

Adds additional container image scanning job using Anchore Grype

* Add missing QEMU setup to Docker workflows (TELDEVOPS-374)

It was noticed that OSS repositories didn't seem to be producing OSS
images correctly, on digging it was spotted that these were missing the
necessary QEMU setup step present in the internal versions of this
workflow.  Adding this step should allow OSS repositories to generate
valid multi-platform images.

* Ping referenced workflows back to main

Signed-off-by: Rob Vesse <rvesse@dotnetrdf.org>

* feat: add anchore scans to workflows

* Enable Trivy caching (CORE-455)

Every workflow that uses Trivy is modified to enable Trivy DB Caching
which should reduce the frequency of rate limiting errors seen when
trying to download the Trivy DB files

* Auto-create Maven Release PRs

Adds a new step to the Maven GitHub Release workflow so that on
successful release it will create a Release PR that completes the
release process.  This avoids the need for developers to manually
open said PRs and reduces the chance that these get missed.

Note no reviewers are assigned, rather we'll rely on default reviewers
which are populated from CODEOWNERS if present

* reverted changes to python workflows (#57)

* fix: grype erroring with invalid sbom

* updated anchore scan version (#59)

* Make opening Maven Release PR not break build

As seen in SCG build if the developer is proactive in creating the
Release PR, or it already exists for some other reason, then this breaks
the build and causes artefacts to not be correctly published.  Make the
Release PR step continue-on-error: true so that failures in that step
are ignored.

* fix(trivyignore): remove

* fix: remove non-main branch

* fix(sbom name): [TELFE-518]remove ".."

---------

Signed-off-by: Rob Vesse <rob.vesse@telicent.io>
Signed-off-by: Rob Vesse <rvesse@dotnetrdf.org>
Co-authored-by: Rob Vesse <rob.vesse@telicent.io>
Co-authored-by: Greg-teli100 <143502364+Greg-teli100@users.noreply.github.com>
Co-authored-by: Andreas Bester <andreas.bester@telicent.io>
Co-authored-by: Georgi <140396136+gnikolov95@users.noreply.github.com>
Co-authored-by: Daniel Davies <110399083+danieldaviestelicent@users.noreply.github.com>
Co-authored-by: Paul Gallagher <132362215+TelicentPaul@users.noreply.github.com>
Co-authored-by: ACoolmanTelicent <a.coleman@telicent.io>